add validate on payment_coef

+++ db/migrate/20201110095904_add_coefficient_to_companies.rb
@@ -1,5 +1,5 @@
 class AddCoefficientToCompanies < ActiveRecord::Migration[6.0]
   def change
-    add_column :companies, :payment_coefficient, :float
+    add_column :companies, :payment_coefficient, :decimal
   end
 end
